fix: satisfy fallow command contract audit
